+++ scalingomock/appsservice_mock.go
@@ -109,6 +109,21 @@ func (mr *MockAppsServiceMockRecorder) AppsList() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AppsList", reflect.TypeOf((*MockAppsService)(nil).AppsList))
 }
 
+// AppsPs mocks base method.
+func (m *MockAppsService) AppsPs(arg0 string) ([]scalingo.Container, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "AppsPs", arg0)
+	ret0, _ := ret[0].([]scalingo.Container)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// AppsPs indicates an expected call of AppsPs.
+func (mr *MockAppsServiceMockRecorder) AppsPs(arg0 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AppsPs", reflect.TypeOf((*MockAppsService)(nil).AppsPs), arg0)
+}
+
 // AppsRename mocks base method.
 func (m *MockAppsService) AppsRename(arg0, arg1 string) (*scalingo.App, error) {
 	m.ctrl.T.Helper()